Testing: Moves failure engine to a generic
